Initial porting of Atom ISP module from Kernel 3.14 to 4.4
This patch enables the basic functionality of Atom ISP and ov7251 sensor

Known issues:
   Intermittently observing GREEN patch when images are captured from
   ov7251 sensor. This would be addressed with a different Jira.

Test results:

- The ov7251 image capures works fine with know issue of Green patch in few frames.
- The kernel crash has been resolved and it was happening due to a variable check
  in kernel 4.4 which has been made mandatory.

In addition, execute checkpatch.pl to resolve Errors from this patch upgrade, as well
as few from the original patch too.

This is intended to unblock the kernel 4.4 work for ICG and NDG teams to enable further
use cases and features. Need to work with ICG team to replace the workaround chosen in
this patch before sent out of Intel.

Signed-off-by: Ghouse Adoni <ghouse.adoni.mohammed@intel.com>