spi: added spi userspace support by enabling device chip select
For atom (including Baytrail and Braswell) the driver starts
DeviceSelection from 1 instead of 0. This fails to enumerate
because ACPI DeviceSelection of 1 is greater than number of
chip selects the driver supports [1].

This series adds a new hook to struct spi_master ->fw_translate_cs()
that allows a driver to translate the chip select number from
firmware to the numbering scheme expected by the Linux SPI core
and implement that for both Baytrail and Braswell.

Source: Back ported from kernel.org

Signed-off-by: Mandeep Bansal <mandeep.bansal@intel.com>