usb otg: add whiskey cove and phy driver to enable otg flow
Adds Whiskey Cove PMIC driver and phy mux implementation to enable
USB OTG detection, charger and temp snippets have been removed
along with portions of the original code to implement basic
features only.

This code is not intended to be submitted to LKML, cleanup and
adaptation is needed for latest kernels. In future, this patch
should be reverted in favor of a driver that has been reviewed
and accepted in LKML.

As such this contains multiple logical blocks to keep changes
contained in a single patch.

Signed-off-by: Omar Ramirez Luna <omar.r.ramirez.luna@intel.com>