jam-stapl: recipe for binary inclusion on aero platform
jam-stapl is a tool released by Altera to interpret and flash JAM
files into an FPGA. Aero platform uses MAX10 FPGA and it is compatible
with jam files. This is a 32-bit static compilation for Aero platform.

Original JAM source does not work well on 64-bit, we need to cross
compile for 32-bit and to avoid including 32-bit librares into the
image we link statically to them.

Once the code is made 64-bit friendly we can remove this dependency.

Signed-off-by: Omar Ramirez Luna <omar.r.ramirez.luna@intel.com>
Signed-off-by: Ghouse Adoni <ghouse.adoni.mohammed@intel.com>