aero: add platform boot configurations
Adding boot option to have display and serial console logging. Also now
the promt to do the installation is going to be in the serial console
instead of the display, avoiding the dependency of have to connect a
monitor in order to do the OS installation.

Sets grub timeout to 3 seconds

Signed-off-by: Israel Cepeda <israel.a.cepeda.lopez@intel.com>