recipe: initrd: change to debug console
OS installer uses S1 console that is available in power connector, that
is fine in order to make the OS installation with nothing else plug in
to that connector (or any other connector with UART1 routed). But the
purpose for S1 is to have a flight controller or something else, not for
debugging. So this patch is going to change the boot console from S1 to
S0 in the grub boot option after the OS is install into eMMC.

If user or developers want to get logs and console into this connector
again, they have to make sure there isn't anything else connected or
enabled (not in HiZ state) and modify the grub.cfg and/or inittab in the
OS and reboot.

Signed-off-by: Israel Cepeda <israel.a.cepeda.lopez@intel.com>