Include FPGA firmware to recover from a bad firmware on AeroFC
Load this FPGA firmware than power off and power on the board, now the
AeroFC will stay on bootloader so you can run aerofc_update.sh and
restore to a working AeroFC firmware.
After that you need to go back to the aero_RTF_kit_fpga.jam to be able
to exit the bootloader and power off and power on the board to boot
on the restored firmware.

This binary was build using the code from
https://github.com/intel-aero/intel-aero-fpga over the commit:
Update Aero fpga sample project 024976f65f6dcb6e920583cc9852d03fabe437e5
using Intel Quartus Prime 16.1 just the output value of
BOOTLOADER_FORCE_PIN register was change to one.