Fix librealsense_v1.12.1 uvc patch to work with Intel Aero
The required kernel patch for librealsense v1.12.1 on krogoth couldn't
be applied directly on the kernel version currently being used by
Intel Aero (4.4-rc3 + custom patches).

This commit replaces the original patch with one that applies the same
changes on the driver but on the needed kernel version. For this to
properly work, its also necessary to bump the priority of this layer
over 'meta-intel-realsense'.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>