librealsense: Install missing R200 config files
For R200 cameras, there are two "config" files at needs to be specifically
installed under ${prefix}/local/bin that aren't being addressed by the
recipe on 'meta-intel-realsense'. Typically those files are installed
via a custom script provided by the lib for users of R200 cameras,
but this script won't work on Yocto environment.

This patch install the missing files and may be dropped once a fix hit
the upstream repository.

This should fix the errors displayed by dmesg when librealsense is used
on the drone.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>