intel-aero-image: change console to ttyS0
Using ttyS1 for console on the RTF is problematic since it's the UART that's
connected to the flight controller. mavlink-router uses this UART and having
the console on it make it unusable. In older versions of the PX4 the 1500000
even triggers the sysrq sequence making the live image unusable.

Thanks to Ederson for helping root causing the issue.