linux: disable some kernel configs
This brings it back to what it was before. Broxton's PMIC needs to be reverted
since it breaks the build and there's no config option to deselect it.

We are now just passing a defconfig since the config from meta-intel is
bringing lots of modules we don't want to add for now. We can revisit this later.