conf: set meta-intel-realsense priority
Ideally meta-intel-realsense would set its priority with a "?=" so we would be
able to set it in the bblayers.conf file. However we need to deal with it for
now, so we reorder the layers to be able to override the priority.

We need to set the priority because now we don't use the kernel URL from meta/,
but rather give our own to skip the cache. meta-intel-realsense needs kernel
patches so they need to be applied after our layer, not before.