intel-aero-repo: bump version to avoid build error
It currently fails incremental build since we are changing the version and
bitbake tries to get the license from the wrong dir. However now the version
for this package is not synced with the distro version anymore. We could just
bump PR variable, but let's bump version to 3, just to make sure it's not
mistaken by the distro version.