[CSD] : Camera def file for Intel RS RGB camera
For Camera control feature, an xml file with description of parameters
that the camera supports need to be hosted over http. The address of the
file will be read by the camera streaming daemon from conf file and sent to
QGC in a mavlink message. QGC will then download the xml file over http